add teachers update/create

+++ database/migrations/2023_10_11_095948_create_teachers_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->text('bio')->nullable();
             $table->string('cv', 2048)->nullable();
             $table->string('photo', 2048)->nullable();
-            $table->unsignedInteger('phone');
+            $table->unsignedBigInteger('phone');
             $table->string('service');
             $table->timestamps();
         });
